Use a ReadOnlySet to be able to quickly determine if the dispatcher supports an event type.
