+++ src/Qowaiv.DomainModel/Collections/ReadOnlySet.cs
@@ -1,0 +1,48 @@
+﻿using Qowaiv.DomainModel.Diagnostics;
+using System.Collections;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Diagnostics.Contracts;
+
+namespace Qowaiv.DomainModel.Collections
+{
+    /// <summary>Represents a read-only set of items.</summary>
+    /// <typeparam name="T">
+    /// The type of items.
+    /// </typeparam>
+    [DebuggerDisplay("Count: {Count}")]
+    [DebuggerTypeProxy(typeof(CollectionDebugView))]
+    public class ReadOnlySet<T> : IEnumerable<T>
+    {
+        [DebuggerBrowsable(DebuggerBrowsableState.Never)]
+        private readonly HashSet<T> collection = new();
+
+        /// <summary>Creates a new instance of the <see cref="ReadOnlySet{T}"/> class.</summary>
+        /// <param name="items">
+        /// The items it contains.
+        /// </param>
+        public ReadOnlySet(IEnumerable<T> items)
+        {
+            Guard.NotNull(items, nameof(items));
+            foreach(var item in items)
+            {
+                collection.Add(item);
+            }
+        }
+
+        /// <summary>Gets the number of elements that are contained in a set.</summary>
+        public int Count => collection.Count;
+
+        /// <summary>Determines whether the set contains the item.</summary>
+        [Pure]
+        public bool Contains(T item) => collection.Contains(item);
+
+        /// <inheritdoc />
+        [Pure]
+        public IEnumerator<T> GetEnumerator() => collection.GetEnumerator();
+
+        /// <inheritdoc />
+        [Pure]
+        IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
+    }
+}
