+++ src/Qowaiv.DomainModel/Dynamic/DynamicEventDispatcher.cs
@@ -1,130 +1,23 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using Qowaiv.DomainModel.Collections;
+using System;
 using System.Diagnostics.Contracts;
 using System.Dynamic;
-using System.Linq;
-using System.Linq.Expressions;
-using System.Reflection;
 
 namespace Qowaiv.DomainModel.Dynamic
 {
     /// <summary>Factory for creating <see cref="DynamicEventDispatcher{TDispatcher}"/>'s.</summary>
-    public abstract class DynamicEventDispatcher : DynamicObject
+    public abstract class DynamicEventDispatcher: DynamicObject
     {
         /// <summary>Gets the supported event types.</summary>
-        public abstract IReadOnlyCollection<Type> SupportedEventTypes { get; }
+        public abstract ReadOnlySet<Type> SupportedEventTypes { get; }
 
         /// <summary>Creates a new instance of the <see cref="DynamicEventDispatcher{TDispatcher}"/> class.</summary>
+        [Pure]
         public static DynamicEventDispatcher New(object dispatcher)
         {
             Guard.NotNull(dispatcher, nameof(dispatcher));
             var type = typeof(DynamicEventDispatcher<>).MakeGenericType(dispatcher.GetType());
             return (DynamicEventDispatcher)Activator.CreateInstance(type, dispatcher);
         }
-    }
-
-    /// <summary>A dynamic event dispatcher, is a extremely limited dynamic object
-    /// that is capable of invoking instance methods with the signature When(@event).
-    /// </summary>
-    /// <typeparam name="TDispatcher">
-    /// The type of the non dynamic event dispatcher.
-    /// </typeparam>
-    /// <remarks>
-    /// The constraints on the method:
-    /// * Name of the method is 'When'
-    /// * Binding is on instance (both public and non-public)
-    /// * One parameter with a type that is/could be an event.
-    /// * Return type is ignored.
-    ///
-    /// It caches the available methods per type.
-    /// </remarks>
-    public class DynamicEventDispatcher<TDispatcher> : DynamicEventDispatcher
-        where TDispatcher : class
-    {
-        /// <summary>Initializes a new instance of the <see cref="DynamicEventDispatcher{TDispatcher}"/> class..</summary>
-        public DynamicEventDispatcher(TDispatcher dispatcher)
-            => this.dispatcher = Guard.NotNull(dispatcher, nameof(dispatcher));
-
-        private readonly TDispatcher dispatcher;
-
-        /// <summary>Tries to invoke a (void) When(@event) method.</summary>
-        /// <exception cref="EventTypeNotSupported">
-        /// If the invoke call was on (void) When(@event) but the type was not available.
-        /// </exception>
-        public override bool TryInvokeMember(InvokeMemberBinder binder, object[] args, out object result)
-            => binder?.Name == nameof(When) && args?.Length == 1 && !(args[0] is null)
-            ? When(args[0], out result)
-            : base.TryInvokeMember(binder, args, out result);
-
-        /// <summary>Gets the supported event types.</summary>
-        public sealed override IReadOnlyCollection<Type> SupportedEventTypes => Lookup.Keys;
-
-        /// <summary>Invokes the When(@event) method.</summary>
-        private bool When(object @event, out object result)
-        {
-            var eventType = @event.GetType();
-            if (Lookup.TryGetValue(eventType, out var when))
-            {
-                result = null;
-                when(dispatcher, @event);
-                return true;
-            }
-            throw new EventTypeNotSupported(eventType, typeof(TDispatcher));
-        }
-
-        /// <summary>Initializes all When(@event) methods.</summary>
-        [Pure]
-        private static Dictionary<Type, Action<TDispatcher, object>> Init()
-        {
-            var lookup = new Dictionary<Type, Action<TDispatcher, object>>();
-
-            foreach (var method in WhenMethods())
-            {
-                var parameters = method.GetParameters();
-                if (parameters.Length == 1)
-                {
-                    var parameterType = parameters[0].ParameterType;
-
-                    // Leave out object itself, primitives and enumerations.
-                    if (parameterType != typeof(object) &&
-                        !parameterType.IsPrimitive &&
-                        !parameterType.IsEnum)
-                    {
-                        lookup[parameterType] = Compile(method, parameterType);
-                    }
-                }
-            }
-            return lookup;
-        }
-
-        /// <summary>Uses reflection to resolve `When` methods to use,
-        /// both non-public (mostly internal) and public.
-        /// </summary>
-        [Pure]
-        private static IEnumerable<MethodInfo> WhenMethods()
-            => typeof(TDispatcher)
-            .GetMethods(MethodSignature)
-            .Where(method => method.Name == nameof(When));
-
-        [Pure]
-        private static Action<TDispatcher, object> Compile(MethodInfo method, Type eventType)
-        {
-            var dispatcherParam = Expression.Parameter(typeof(TDispatcher), "dispatcher");
-            var @event = Expression.Parameter(typeof(object), "e");
-
-            var typedEvent = Expression.Convert(@event, eventType);
-            var body = Expression.Call(dispatcherParam, method, typedEvent);
-
-            var expression = Expression.Lambda<Action<TDispatcher, object>>(body, dispatcherParam, @event);
-
-            return expression.Compile();
-        }
-
-        private static readonly Dictionary<Type, Action<TDispatcher, object>> Lookup = Init();
-
-#pragma warning disable S3011 // Reflection should not be used to increase accessibility of classes, methods, or fields
-        /// <remarks>We explicitly </remarks>
-        private const BindingFlags MethodSignature = BindingFlags.Public | BindingFlags.NonPublic| BindingFlags.Instance;
-#pragma warning restore S3011 // Reflection should not be used to increase accessibility of classes, methods, or fields
     }
 }
